Add report, setup-client. Changed config layout.
The hooks need to be updated to match the new layout. Keys are only
added to the branch, not the remote.

For clients, we modify the remote to add/change "url", and the branch to
add "assure-key". For publishers, we modify the remote to set
"pushurl" (to the base url) and "push" (to include refs/notes/commits),
and the branch to include "assure-sign-key".